Refatorando a página de detalhes
