+++ src/components/index.js
@@ -1,5 +1,8 @@
 export { default as Card } from './Card';
+export { default as DetailsHeader } from './DetailsHeader';
+export { default as FavoriteButton } from './FavoriteButton';
 export { default as Filters } from './Filters';
 export { default as Footer } from './Footer';
 export { default as Header } from './Header';
 export { default as SearchBar } from './SearchBar';
+export { default as ShareButton } from './ShareButton';
