+++ src/pages/RecipeDetails/index.js
@@ -1,18 +1,12 @@
-/* eslint-disable max-lines-per-function */
 import React, { useEffect, useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
-import copy from 'clipboard-copy';
 import RequestData from '../../services/RequestAPI';
-import shareIcon from '../../images/shareIcon.svg';
-import blackHeartIcon from '../../images/blackHeartIcon.svg';
-import whiteHeartIcon from '../../images/whiteHeartIcon.svg';
+import { DetailsHeader } from '../../components';
 import './style.css';
-import { getStorage, setStorage } from '../../services/localStorage';
+import { getStorage } from '../../services/localStorage';
 
 function RecipeDetails() {
   const [details, setDetails] = useState();
-  const [shared, setShared] = useState(false);
-  const [favorite, setFavorite] = useState(false);
   const [recomendation, setRecomendation] = useState();
   const [inProgress, setInProgress] = useState(false);
   const { category, idReceita } = useParams();
@@ -78,98 +72,12 @@ function RecipeDetails() {
     );
   }
 
-  useEffect(() => {
-    const localFavorite = getStorage('favoriteRecipes');
-    if (localFavorite) {
-      const bool = localFavorite.find((item) => item.id === idReceita);
-      setFavorite(bool);
-    }
-  }, [idReceita]);
-
-  function handleFavorite() {
-    let localFavorite = getStorage('favoriteRecipes');
-    if (localFavorite) {
-      const bool = localFavorite.find((item) => item.id === idReceita);
-      setFavorite(!bool);
-      if (bool) {
-        const newFavorite = localFavorite.filter((item) => item.id !== idReceita);
-        setStorage('favoriteRecipes', newFavorite);
-        return;
-      }
-    } else {
-      localFavorite = [];
-      setFavorite(true);
-    }
-    const obj = {
-      id: idReceita,
-      type: category,
-      area: details.strArea || '',
-      category: details.strAlcoholic || details.strCategory,
-      alcoholicOrNot: details.strAlcoholic || '',
-      name: details.strDrink || details.strMeal,
-      image: details.strDrinkThumb || details.strMealThumb,
-    };
-    localFavorite.push(obj);
-    setStorage('favoriteRecipes', localFavorite);
-  }
-
-  function sharedButton() {
-    setShared(true);
-    const time = 3000;
-    setTimeout(() => setShared(false), time);
-  }
-
   return (
     <div>
       {!details && 'Loading...'}
       {details && (
         <div>
-          <img
-            data-testid="recipe-photo"
-            src={ details.strDrinkThumb || details.strMealThumb }
-            alt="drink"
-          />
-          <h3 data-testid="recipe-title">
-            {details.strDrink || details.strMeal}
-          </h3>
-          <button
-            type="button"
-            data-testid="share-btn"
-            onClick={ () => {
-              sharedButton();
-              copy(window.location.href);
-            } }
-          >
-            <img
-              src={ shareIcon }
-              alt="Compartilhar"
-            />
-            {(shared) && <span>Link copiado!</span>}
-          </button>
-          <button
-            type="button"
-            onClick={ () => handleFavorite() }
-          >
-            {
-              !favorite
-                ? (
-                  <img
-                    data-testid="favorite-btn"
-                    src={ whiteHeartIcon }
-                    alt="favoritar"
-                  />
-                ) : (
-                  <img
-                    data-testid="favorite-btn"
-                    src={ blackHeartIcon }
-                    alt="favoritar"
-                  />
-                )
-            }
-          </button>
-          <p data-testid="recipe-category">
-            {details.strAlcoholic || details.strCategory}
-          </p>
+          <DetailsHeader details={ details } />
           {takeIngredients().map((ingredients, index) => (
             <p
               data-testid={ `${index}-ingredient-name-and-measure` }
